# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: FixingApp Backend API
  description: |
    Backend MVP API for connecting employers with workers.
    
    This API provides endpoints for:
    - Authentication and user management
    - Job posting and management
    - Job applications
    - Worker verification
    - Reviews and ratings
    - Complaints handling
    - Admin functions
    
    **Authentication**: Most endpoints require a JWT Bearer token in the Authorization header.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fixingapp.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3000/api
    description: API base path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        message:
          type: string
          description: Detailed error message
      example:
        error: "Not Found"
        message: "Route GET /api/invalid not found"
...
    Job:
      type: object
      properties:
        id:
          type: integer
          example: 1
        employerId:
          type: integer
          example: 1
        title:
          type: string
          example: "Fix leaking pipe"
        description:
          type: string
          example: "Need to fix a leaking pipe in kitchen"
        price:
          type: integer
          format: int64
          example: 150000
        address:
          type: string
          example: "123 Main St, City"
        requiredSkill:
          type: string
          example: "Plumbing"
        status:
          type: string
          enum: [CHUA_LAM, DANG_BAN_GIAO, DA_XONG]
          example: "CHUA_LAM"
        acceptedWorkerId:
          type: integer
          nullable: true
          example: null
        createdAt:
          type: integer
          format: int64
          example: 1702719041000
        updatedAt:
          type: integer
          format: int64
          example: 1702719041000
paths:
...
  /api/jobs:
    get:
      tags:
        - Jobs
      summary: List available jobs
      description: Get a list of open jobs (CHUA_LAM, DANG_BAN_GIAO status)
      operationId: listJobs
      parameters:
        - name: keyword
          in: query
          description: Search keyword
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category/skill
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: integer
        - name: maxPrice
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'

    post:
      tags:
        - Jobs
      summary: Create a new job
      description: Create a new job posting (Employer only)
      operationId: createJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - price
                - address
              properties:
                title:
                  type: string
                  example: "Fix leaking pipe"
                description:
                  type: string
                  example: "Need to fix a leaking pipe in kitchen"
                price:
                  type: integer
                  format: int64
                  example: 150000
                address:
                  type: string
                  example: "123 Main St, City"
                requiredSkill:
                  type: string
                  example: "Plumbing"
                images:
                  type: array
                  items:
                    type: string
                  example: ["/static/pipe1.jpg"]
            example:
              title: "Fix leaking pipe"
              description: "Need to fix a leaking pipe in kitchen"
              price: 150000
              address: "123 Main St, City"
              requiredSkill: "Plumbing"
              images: ["/static/pipe1.jpg"]
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/workers/certificates:
    post:
      tags:
        - Workers
      summary: Upload certificate
      description: Upload a professional certificate (Worker only)
      operationId: uploadCertificate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - imageUrl
              properties:
                imageUrl:
                  type: string
                  example: "/static/IELTS_7.5.jpg"
            example:
              imageUrl: "/static/IELTS_7.5.jpg"
      responses:
        '201':
          description: Certificate uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  workerId:
                    type: integer
                  imageUrl:
                    type: string
                  status:
                    type: string
                    enum: [PENDING]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/complaints:
    post:
      tags:
        - Complaints
      summary: File a complaint
      description: Submit a complaint about a job
      operationId: fileComplaint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jobId
                - reason
              properties:
                jobId:
                  type: integer
                  example: 1
                reason:
                  type: string
                  example: "Worker did not show up on time"
                evidences:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [IMAGE, LOG]
                      url:
                        type: string
            example:
              jobId: 1
              reason: "Worker did not show up on time"
              evidences:
                - type: "IMAGE"
                  url: "/static/evidence1.jpg"
      responses:
        '201':
          description: Complaint filed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  jobId:
                    type: integer
                  reason:
                    type: string
                  status:
                    type: string
                    enum: [PENDING]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/jobs/pending:
    get:
      tags:
        - Admin
      summary: Get pending jobs
      description: Get all jobs pending approval (Admin only)
      operationId: getPendingJobs
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of pending jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/jobs/{jobId}/approve:
    post:
      tags:
        - Admin
      summary: Approve a job
      description: Approve a job posting (Admin only)
      operationId: approveJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Job approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/jobs/{jobId}/reject:
    post:
      tags:
        - Admin
      summary: Reject a job
      description: Reject a job posting (Admin only)
      operationId: rejectJob
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "Job description violates guidelines"
      responses:
        '200':
          description: Job rejected successfully
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/certificates/pending:
    get:
      tags:
        - Admin
      summary: Get pending certificates
      description: Get all certificates pending verification (Admin only)
      operationId: getPendingCertificates
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of pending certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    workerId:
                      type: integer
                    imageUrl:
                      type: string
                    status:
                      type: string
                      enum: [PENDING]
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/certificates/{certId}/approve:
    post:
      tags:
        - Admin
      summary: Approve a certificate
      description: Approve a worker certificate (Admin only)
      operationId: approveCertificate
      security:
        - bearerAuth: []
      parameters:
        - name: certId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Certificate approved successfully
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/certificates/{certId}/reject:
    post:
      tags:
        - Admin
      summary: Reject a certificate
      description: Reject a worker certificate (Admin only)
      operationId: rejectCertificate
      security:
        - bearerAuth: []
      parameters:
        - name: certId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  example: "Certificate image is not clear"
      responses:
        '200':
          description: Certificate rejected successfully
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/admin/complaints:
    get:
      tags:
        - Admin
      summary: Get all complaints
      description: Get all complaints (Admin only)
      operationId: getAllComplaints
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '403':
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...